requirements.txt
streamlit>=1.20.0
boto3>=1.26.0
pandas>=2.0.0
yfinance>=0.2.0
chromadb>=0.4.0
google-search-results>=2.4.0
python-dotenv>=1.0.0
